requirements.txt
torch>=2.0.0
numpy>=1.20.0
pyyaml>=6.0
wandb>=0.15.0
matplotlib>=3.5.0
tqdm>=4.65.0
click>=8.1.0
optuna>=3.0.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
isort>=5.10.0
mypy>=1.0.0
ruff>=0.0.200